完成了 流程控制（选择、循环结构） 实验作业:

+++ 181001 流程控制-选择与循环结构/test.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-    double n = 1;
+    int n = 1;
     double sum = 0.0;
     int lmt;
     /**
